Chapters modifications
An error now occur when the admin try to publish a chapter with a negative or a null number.
An error now occur when the admin try to publish a chapter with an already used number.
Chapters now theyre numbers displayed on homepage and chapter page.
Chapters are now created when the admin save them for the first time.
All dates problems are now fixed